Moved create schema out of root.rs
