+++ backend/src/schema.rs
@@ -1,0 +1,16 @@
+mod product;
+mod root;
+mod user;
+
+use self::root::{MutationRoot, QueryRoot};
+use juniper::{EmptySubscription, RootNode};
+
+pub struct Context {}
+
+impl juniper::Context for Context {}
+
+pub type Schema = RootNode<'static, QueryRoot, MutationRoot, EmptySubscription<Context>>;
+
+pub fn create_schema() -> Schema {
+    Schema::new(QueryRoot, MutationRoot, EmptySubscription::new())
+}
