--- backend/src/schemas/mod.rs
@@ -1,3 +1,0 @@
-pub mod product;
-pub mod root;
-pub mod user;
